Video transform show progress, make cancel work, add transcode to mp4
Previously, progress was only shown when there were multiple files by informing the user which file was being transcoded. As transcoding video can take a long time, that was suboptimal. Now we attach to ffmpeg pipe and show percentage of the current video being transcoded.

Also, cancelling now works and cleans up after itself.

Option to specifically transcode to mp4 was added - extensions are handled in such a case.

The code is somewhat awful, but it seems to work.