Fix variable name and also handling of spaces
ATTACHMENTS is always one file, so have it in singular. It is reused later, but is always reset, so that should not be a problem.

Previously, files containing spaces and starting with a hyphen broke the temporary directory, this fixes that. Weirdly, using -- from previous commit right after magick is ok, because it only applies to the very next file and -resize is again interpretted as a command.